pyproject.toml
[tool.poetry]
name = "tap-prometheus"
version = "0.0.1"
description = "Singer tap that extracts data from Prometheus using queries provided in config and produces JSON-formatted data following the [Singer spec](https://github.com/singer-io/getting-started/blob/master/docs/SPEC.md)."
authors = ["Signal AI"]
repository = "https://github.com/signal-ai/tap-prometheus"
homepage = "https://github.com/signal-ai/tap-prometheus"
packages = [ { include="tap_prometheus" } ]

[tool.poetry.scripts]
tap-prometheus = "tap_prometheus:main"

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0.0"
singer-python = "5.12.2"
prometheus-api-client = "^0.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-httpserver = "^1.0.6"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
